separate concerns between abstract dialog class  and osd dialog class
